refactor: components/common에 공용 컴포넌트 폴더 구조 개선
- 각 용도별로 분류
