+++ frontend/src/components/common/Input/SearchBar.tsx
