+++ frontend/src/components/mypage/ButtonSection.tsx
@@ -1,4 +1,4 @@
-import Button from "@components/common/Button";
+import Button from "@/components/common/Button/DefaultButton";
 
 interface ButtonSectionProps {
   closeModal: () => void;
