+++ frontend/src/components/questions/create/QuestionForm/CategorySection/index.tsx
@@ -1,7 +1,7 @@
-import SelectTitle from "@/components/common/SelectTitle";
+import SelectTitle from "@/components/common/Text/SelectTitle";
 import { options } from "./data";
 import useQuestionFormStore from "@/stores/useQuestionFormStore";
-import Select from "@/components/common/Select";
+import CategorySelect from "@/components/common/Select/CategorySelect";
 
 const CategorySection = () => {
   const { category, setCategory } = useQuestionFormStore();
@@ -10,7 +10,7 @@ const CategorySection = () => {
     <div>
       <SelectTitle title="카테고리" />
       <div className="h-11">
-        <Select
+        <CategorySelect
           defaultText="카테고리를 선택해주세요"
           options={options}
           value={category}
