+++ frontend/src/components/questions/create/QuestionForm/QuestionInputSection/index.tsx
@@ -1,4 +1,4 @@
-import SelectTitle from "@/components/common/SelectTitle";
+import SelectTitle from "@/components/common/Text/SelectTitle";
 import QuestionInput from "./QuestionInput";
 import QuestionList from "./QuestionList";
 import useQuestionFormStore from "@/stores/useQuestionFormStore";
