+++ frontend/src/components/questions/create/QuestionForm/TitleSection/index.tsx
@@ -1,5 +1,5 @@
-import SelectTitle from "@/components/common/SelectTitle";
-import TitleInput from "@/components/common/TitleInput";
+import SelectTitle from "@/components/common/Text/SelectTitle";
+import TitleInput from "@/components/common/Input/TitleInput";
 import useQuestionFormStore from "@/stores/useQuestionFormStore";
 
 const TitleSection = () => {
