+++ frontend/src/components/sessions/create/SessionForm/AccessSection/index.tsx
@@ -1,6 +1,6 @@
 import useSessionFormStore from "@/stores/useSessionFormStore";
-import SelectTitle from "@/components/common/SelectTitle";
-import AccessButton from "@/components/common/AccessButton";
+import SelectTitle from "@/components/common/Text/SelectTitle";
+import AccessButton from "@/components/common/Button/AccessButton";
 
 const AccessSection = () => {
   const { access, setAccess } = useSessionFormStore();
