+++ frontend/src/components/sessions/create/SessionForm/ParticipantSection/index.tsx
@@ -1,5 +1,5 @@
 import ParticipantButton from "./ParticpantButton";
-import SelectTitle from "@/components/common/SelectTitle";
+import SelectTitle from "@/components/common/Text/SelectTitle";
 import useSessionFormStore from "@/stores/useSessionFormStore";
 const MEMBER_COUNTS = [1, 2, 3, 4, 5] as const;
