+++ frontend/src/main.tsx
@@ -2,7 +2,7 @@ import { createRoot } from "react-dom/client";
 import "./index.css";
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import { routes } from "./routes.tsx";
-import ToastProvider from "./components/common/ToastProvider.tsx";
+import ToastProvider from "./components/common/Toast/ToastProvider.tsx";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
